Mark infrastructure code as internal
