+++ src/Evlog.Infrastructure/DataModels/UserDM.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace Evlog.Infrastructure.DataModels
 {
-    public class UserDM
+    internal class UserDM
     {
         public int Id { get; set; }
         public string Email { get; set; }
