+++ src/Evlog.Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -1,3 +1,9 @@
+using Evlog.Domain.EventAggregate.Commands;
+using Evlog.Domain.EventAggregate.Queries;
+using Evlog.Domain.UserAggregate.Commands;
+using Evlog.Domain.UserAggregate.Queries;
+using Evlog.Infrastructure.Commands;
+using Evlog.Infrastructure.Queries;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,6 +24,25 @@ namespace Evlog.Infrastructure.Extensions
             });
 
             return services;
+        }
+
+        public static void AddEvlogQueries(this IServiceCollection services)
+        {
+            // Add EventPost queries
+            services.AddTransient<IAllEventsQuery, AllEventsQuery>();
+            services.AddTransient<IPastEventsQuery, PastEventsQuery>();
+            services.AddTransient<IUpcomingEventsQuery, UpcomingEventsQuery>();
+            services.AddTransient<IEventQuery, EventQuery>();
+
+            // Add User queries
+            services.AddTransient<IUserQuery, UserQuery>();
+            services.AddTransient<IUserExistsQuery, UserExistsQuery>();
+        }
+
+        public static void AddEvlogCommands(this IServiceCollection services)
+        {
+            services.AddTransient<IRegisterUserCommand, RegisterUserCommand>();
+            services.AddTransient<ICreateUserCommand, CreateUserCommand>();
         }
     }
 }
